+++ api/src/main/java/me/ramos/api/controller/QueryPlayerController.java
@@ -3,7 +3,7 @@ package me.ramos.api.controller;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import me.ramos.commons.dto.QueryPlayerResponse;
-import me.ramos.commons.service.QueryPlayerService;
+import me.ramos.commons.service.impl.QueryPlayerService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
