+++ api/src/main/java/me/ramos/api/controller/QueryTeamController.java
@@ -3,7 +3,7 @@ package me.ramos.api.controller;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import me.ramos.commons.domain.model.Team;
-import me.ramos.commons.service.QueryTeamService;
+import me.ramos.commons.service.impl.QueryTeamService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
