+++ api/src/main/java/me/ramos/api/controller/TestApiController.java
@@ -1,7 +1,7 @@
 package me.ramos.api.controller;
 
 import lombok.RequiredArgsConstructor;
-import me.ramos.commons.service.TestCommonService;
+import me.ramos.commons.service.impl.TestCommonService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
