+++ commons/src/main/java/me/ramos/commons/service/impl/QueryTeamService.java
@@ -1,4 +1,4 @@
-package me.ramos.commons.service;
+package me.ramos.commons.service.impl;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
